Small adjustments
