+++ backend/src/main/kotlin/hu/bme/sch/cmsch/component/team/TeamIntroductionReviewController.kt
@@ -4,26 +4,13 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import hu.bme.sch.cmsch.component.login.CmschUser
 import hu.bme.sch.cmsch.controller.admin.OneDeepEntityPage
 import hu.bme.sch.cmsch.controller.admin.calculateSearchSettings
-import hu.bme.sch.cmsch.repository.EntityPageDataSource
 import hu.bme.sch.cmsch.service.*
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
 import org.springframework.core.env.Environment
-import org.springframework.data.jpa.repository.Query
-import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Controller
-import org.springframework.stereotype.Repository
 import org.springframework.transaction.PlatformTransactionManager
 import org.springframework.web.bind.annotation.RequestMapping
 
-
-@Repository
-interface TeamIntroductionRepository :
-    CrudRepository<TeamIntroductionEntity, Int>,
-    EntityPageDataSource<TeamIntroductionEntity, Int> {
-
-    @Query("SELECT t FROM TeamIntroductionEntity t WHERE t.group.id = :groupId ORDER BY t.creationDate DESC")
-    fun findIntroductionsForGroup(groupId: Int): List<TeamIntroductionEntity>
-}
 
 @Controller
 @RequestMapping("/admin/control/team-introductions")
@@ -67,11 +54,11 @@ class TeamIntroductionReviewController(
     adminMenuPriority = 2,
 
     searchSettings = calculateSearchSettings<TeamIntroductionEntity>(false)
-)
-{
+) {
     override fun fetchOverview(user: CmschUser): Iterable<TeamIntroductionEntity> {
         return teamIntroductionRepository.findAll()
-            .groupBy { it.group?.id ?: 0 }
-            .map { entries -> entries.value.maxBy { it.creationDate } }
+            .sortedWith(
+                compareBy<TeamIntroductionEntity> { it.groupName }.thenComparing { it.creationDate }.reversed()
+            )
     }
 }
